[refactor] comment index 에서 export 및 그에 따른 import 구문 수정
